tidy - remove old EventHandlerBuilder
